WIFI_REALTEK to WIFI_RTW, NANOSTACK to features
The name of the WIFI should follow the actual class name, this aligns
the example code as well to match the change done in easy-connect.

Adding NANOSTACK to features as well - this enables the code to compile
even with the .mbedignore and now it also works again (there were issues
prior that ethernet would not work, if nanostack was included).

This should remove the need to have the .mbedignore files now.